Fix updating routes when uploading one
Now, the route is uploaded, and automatically reloaded. I've set a delay of 3s, so when you pressed the close button on the modal the function to upload the routes is triggered in within that interval of time. Thanks to our middlewares, when the routes are loading, a loder is displayed to the user to notify them that the routes are being updated.